Refactor v2 trace conditions
